requirements.txt
groq
scrapy
selenium
webdriver-manager
inquirer
python-dotenv
prompt_toolkit==1.0.14
tavily-python
